# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "POC",
        "version": "1.0.0",
        "description": "API Server",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "Dheeraj Budhiraja",
            "email": "itsdheerajbudhiraja@gmail.com"
        }
    },
    "paths": {
        "/docs/swagger/{docsApiKey}/swagger.json": {
            "get": {
                "summary": "Download JSON of Swagger Docs",
                "tags": [
                    "docs"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "docsApiKey",
                        "description": "DOCS API KEY",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Swagger Docs for REST API endpoints"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/docs/swagger/{docsApiKey}": {
            "get": {
                "summary": "Swagger Docs UI",
                "tags": [
                    "docs"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "docsApiKey",
                        "description": "DOCS API KEY",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Swagger Docs for REST API endpoints"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/docs/typedoc/{docsApiKey}": {
            "get": {
                "summary": "Typescript docs",
                "tags": [
                    "docs"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "docsApiKey",
                        "description": "DOCS API KEY",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Typescript docs"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/docs/test-report/{docsApiKey}": {
            "get": {
                "summary": "Test Report",
                "tags": [
                    "docs"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "docsApiKey",
                        "description": "DOCS API KEY",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test Report"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            },
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
    "servers": [
        {
            "url": "http://localhost:3000/"
        }
    ]
}